Web/edmilson/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyJWT==1.5.0
Werkzeug==2.0.2
